Fixed bug where client could not connect on Win x64 builds; clarified warning when an attempt is made to recompile all assets in release builds

+++ Source/Server/packet.c
@@ -62,7 +62,7 @@ bool packet_recv(Packet *p) {
 
     size_t app_id;
     p->a = header_unpack(p->p, &app_id, &p->ack, &p->time);
-    if(app_id != APP_ID) return false; /* TODO: warn */
+    if((int32_t)app_id != APP_ID) return false; /* TODO: warn */
     return true;
 }
